split component to reused
